+++ client/src/Admin/AdminRosters/AdminRosterDetail.jsx
@@ -53,10 +53,10 @@ class AdminRosterDetail extends Component {
   render() {
     let rosterEmails = (this.state.camp.registrations || [])
       .filter(registration => registration.paid || registration.deposit)
-      .map(registration => registration.user.email);
+      .map(registration => registration.user.primaryContact.email);
     let depositEmails = (this.state.camp.registrations || [])
       .filter(registration => !registration.paid && registration.deposit)
-      .map(registration => registration.user.email);
+      .map(registration => registration.user.primaryContact.email);
     return (
       <div className="admin-roster-detail">
         <div className="card spinner-wrapper">